requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.11.0
click==8.1.7
colorama==0.4.6
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
flake8==6.1.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.0.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
rpds-py==0.13.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
